rename message screen to application screen
